Exclude header from the output csv file
Issue #11: Allow the user to exclude the header from the output csv file

[Problem]
The user might want to exclude the header line from the csv file. This
is useful if appending to a google sheets document.

[Solution]
Allow the user to exclude the header line from the output csv file.
Default is to still include the header.