System to configure input format
Issue #21: Create a system to allow customizing the data format.

[Problem]
Currently, supporting a new bank transaction format requires writing new
code. This is a pretty big operational burden and also requires code
changes if a bank changes its format.

[Solution]
Move the definition of transaction formats to a configuration file so
that the format can be changed with out any code changes.