Args and config file in single interface
Issue #28: Abstract the command line arguments and the config file
behind a single interface.

[Problem]
There are duplicated options in the command line arguments and the
config file, and within the config file. This makes it difficult to know
where to look up an option and the logic for which option overrides
another is not in a single location.

[Solution]
Move the arguments module to a submodule of the config module. Combine
the arguments and the config file objects into a single config object
that provides methods to get the various options, correctly overriding
options where necessary.